# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Board of Issues task management system.",
        "title": "Board Of Issues API",
        "contact": {
            "name": "Nikita Kleymenov",
            "url": "https://t.me/n1ktarchik",
            "email": "klejmenov663@email.com"
        },
        "version": "2.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tasks/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a task by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "\"550e8400-e29b-41d4-a716-446655440000\"",
                        "description": "TASK ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: task with ID ... has been deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_uuid, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_an_owner",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "task_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tasks/{taskId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve full information about a specific task by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task details",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "\"550e8400-e29b-41d4-a716-446655440000\"",
                        "description": "Task UUID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication (format: Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.Task"
                        }
                    },
                    "400": {
                        "description": "Possible: invalid_uuid, bad_request",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "not_a_desk_member",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "task_not_found",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_server_error",
                        "schema": {
                            "$ref": "#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "N1ktarchik_Board_of_issues_internal_core_domain.Task": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "636e856-e12b-56d9-f987-333222561234"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "deadline": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Task description"
                },
                "desk_id": {
                    "type": "string",
                    "example": "832t758-a12g-47y9-i999-123456789098"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "Task name"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "error": {
                    "type": "string",
                    "example": "any string-code of error"
                },
                "message": {
                    "type": "string",
                    "example": "any error message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token in the format: Bearer \u003cJWT_TOKEN\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "IdempotencyKey": {
            "description": "An idempotent key to prevent duplicates (UUID)",
            "type": "apiKey",
            "name": "X-Req-Key",
            "in": "header"
        }
    }
}